requirements.txt
opencv-python
pillow
ultralytics
openai
python-dotenv